+++ src/aux-server/aux-web/aux-projector/App/App.vue
@@ -129,6 +129,8 @@
                 </md-dialog-actions>
             </md-dialog>
 
+            <!-- <loading></loading> -->
+
             <md-dialog-confirm
             :md-active.sync="showConfirmDialog"
             v-bind:md-title="confirmDialogOptions.title"