+++ src/aux-server/aux-web/aux-projector/index.ts
@@ -62,6 +62,7 @@ import App from './App/App';
 import Welcome from './Welcome/Welcome';
 import Home from './Home/Home';
 import AuxDebug from './AuxDebug/AuxDebug';
+import Loading from '../shared/vue-components/Loading/Loading';
 
 // Import the WebXR Polyfill
 import 'webxr-polyfill';
@@ -160,11 +161,16 @@ router.beforeEach((to, from, next) => {
 });
 
 async function start() {
+    const loading = new Vue({
+        render: createEle => createEle(Loading)
+    }).$mount('#loading');
+
     await appManager.initPromise;
     const app = new Vue({
         router,
         render: createEle => createEle(App)
     }).$mount('#app');
+
 }
 
 start();