+++ src/aux-server/CHANGELOG.md
@@ -1,5 +1,12 @@
 # AUX Server Changelog
 
+## V0.4.2
+### Date: TBD
+
+### Changes:
+- Improvements
+  - Added top-level loading screen to Aux Builder and Aux Player.
+  
 ## V0.4.1
 ### Date: 4/05/2019